Add files via upload
This is the video representation of the project . Its about how the project is done  by creating various objects, field and relationships, profile, and users and giving permission access to the objects and profile. Then hoe to create  a record,  view a record and delete the record. Then the last is the report  to create and view the report.
THANK YOU